refactor snapshot service a tad (#1531)
just as wee bit of shuffling. still quite messy

also includes some updates for smart snapshot chunking
