+++ packages/client/src/utils/iterators.ts
