+++ packages/client/src/utils/numbers/index.ts
@@ -1,4 +1,4 @@
 export { parseTokenBalance } from './balances';
-export { numberToHex } from './hex';
+export { numberToHex, uint8ArrayToHexString } from './hex';
 export { calcPercent, calcPercentCompletion } from './percents';
 export { getRateDisplay } from './rates';
