+++ packages/client/src/workers/sync/utils.ts
@@ -1,28 +1,18 @@
 import { JsonRpcProvider } from '@ethersproject/providers';
-import { ComponentValue, Components, EntityID } from '@mud-classic/recs';
+import { Components, EntityID } from '@mud-classic/recs';
 import { abi as WorldAbi } from '@mud-classic/solecs/abi/World.json';
 import { World } from '@mud-classic/solecs/types/ethers-contracts';
-import { awaitPromise, range, to256BitString, unpackTuple } from '@mud-classic/utils';
-import { BigNumber, BytesLike } from 'ethers';
-import { createChannel, createClient } from 'nice-grpc-web';
+import { awaitPromise, range, to256BitString } from '@mud-classic/utils';
+import { BigNumber } from 'ethers';
 import { Observable, concatMap, map, of } from 'rxjs';
 
-import { grpc } from '@improbable-eng/grpc-web';
-import { createDecoder } from 'engine/encoders';
+import { createDecode } from 'engine/encoders';
 import {
   ECSStateReplyV2,
   ECSStateSnapshotServiceClient,
 } from 'engine/types/ecs-snapshot/ecs-snapshot';
-import {
-  BlockResponse,
-  Component,
-  Entity,
-  KamigazeServiceClient,
-  KamigazeServiceDefinition,
-  State,
-} from 'engine/types/kamigaze/kamigaze';
 import { formatComponentID, formatEntityID } from 'engine/utils';
-import { ComponentsSchema } from 'types/ComponentsSchema';
+import { uint8ArrayToHexString } from 'utils/numbers';
 import { ContractConfig } from 'workers/types';
 import { debug as parentDebug } from '../debug';
 import {
@@ -33,216 +23,11 @@ import {
   SystemCallTransaction,
 } from '../types';
 import { fetchEventsInBlockRange } from './blocks';
-import { CacheStore, createCacheStore, storeEvent } from './cache';
+import { CacheStore, createStateCache, storeEvent } from './cache';
 import { createTopics } from './topics';
 
 const debug = parentDebug.extend('syncUtils');
-interface FetchOptions {
-  cacheStore: CacheStore;
-  kamigazeClient: KamigazeServiceClient;
-  decode: ReturnType<typeof createDecode>;
-  numChunks?: number;
-  setPercentage: (percentage: number) => void;
-}
 
-/**
- * KAMIGAZE INTEGRATION
- **/
-
-export function createSnapshotClient(url: string): KamigazeServiceClient {
-  return createClient(KamigazeServiceDefinition, createChannel(url, grpc.WebsocketTransport()));
-}
-
-export async function fetchStateFromKamigaze(
-  cacheStore: CacheStore,
-  kamigazeClient: KamigazeServiceClient,
-  decode: ReturnType<typeof createDecode>,
-  numChunks = 10,
-  setPercentage: (percentage: number) => void
-): Promise<CacheStore> {
-  let currentBlock = cacheStore.lastKamigazeBlock;
-  let initialLoad = currentBlock == 0;
-
-  const options: FetchOptions = { cacheStore, kamigazeClient, decode, numChunks, setPercentage };
-
-  //block
-  let BlockResponse = await kamigazeClient.getStateBlock({});
-  if (cacheStore.kamigazeNonce != BlockResponse.nonce) {
-    console.log('New nonce found, full state load required');
-    options.cacheStore = createCacheStore();
-    initialLoad = true;
-  }
-
-  try {
-    await fetchComponents(options);
-    if (!initialLoad) {
-      await fetchStateRemovals(options);
-    }
-    await fetchStateValues(options);
-    await fetchEntities(options);
-  } catch (error) {
-    console.error('Error:', error);
-    throw error;
-  }
-
-  storeBlock(options.cacheStore, BlockResponse);
-  options.cacheStore.lastKamigazeBlock = BlockResponse.blockNumber;
-  options.cacheStore.kamigazeNonce = BlockResponse.nonce;
-
-  return options.cacheStore;
-}
-
-async function fetchComponents({ cacheStore, kamigazeClient, setPercentage }: FetchOptions) {
-  // remove from the cache any component added by the rpc sync
-  cacheStore.components.splice(cacheStore.lastKamigazeComponent + 1);
-
-  const ComponentsResponse = await kamigazeClient.getComponents({
-    fromIdx: cacheStore.lastKamigazeComponent,
-  });
-
-  storeComponents(cacheStore, ComponentsResponse.components);
-  cacheStore.lastKamigazeComponent = cacheStore.components.length - 1;
-  setPercentage(5);
-}
-
-async function fetchStateRemovals({
-  cacheStore,
-  kamigazeClient,
-  numChunks = 10,
-  setPercentage,
-}: FetchOptions) {
-  let percent = 5;
-  const StateRemovalsResponse = await kamigazeClient.getState({
-    fromBlock: cacheStore.lastKamigazeBlock,
-    numChunks,
-    removals: true,
-  });
-  for await (const responseChunk of StateRemovalsResponse) {
-    removeValues(cacheStore, responseChunk.state);
-    setPercentage(percent);
-    percent += 3;
-  }
-}
-
-async function fetchStateValues({
-  cacheStore,
-  kamigazeClient,
-  decode,
-  numChunks = 10,
-  setPercentage,
-}: FetchOptions) {
-  let percent = 40;
-  const StateValuesResponse = await kamigazeClient.getState({
-    fromBlock: cacheStore.lastKamigazeBlock,
-    numChunks,
-    removals: false,
-  });
-
-  for await (const responseChunk of StateValuesResponse) {
-    storeValues(cacheStore, responseChunk.state, decode);
-    setPercentage(percent);
-    percent += 3;
-  }
-}
-
-async function fetchEntities({
-  cacheStore,
-  kamigazeClient,
-  numChunks = 10,
-  setPercentage,
-}: FetchOptions) {
-  // remove from the cache any entity added by the rpc sync
-  let percent = 75;
-  cacheStore.entities.splice(cacheStore.lastKamigazeEntity + 1);
-
-  const EntitiesResponse = kamigazeClient.getEntities({
-    fromIdx: cacheStore.lastKamigazeEntity,
-    numChunks,
-  });
-
-  for await (const responseChunk of EntitiesResponse) {
-    storeEntities(cacheStore, responseChunk.entities);
-    setPercentage(percent);
-    percent += 3;
-  }
-
-  cacheStore.lastKamigazeEntity = cacheStore.entities.length - 1;
-}
-
-export function storeBlock(cacheStore: CacheStore, block: BlockResponse) {
-  cacheStore.blockNumber = block.blockNumber;
-  console.log('Stored block');
-}
-export function storeComponents(cacheStore: CacheStore, components: Component[]) {
-  if (typeof components === 'undefined') {
-    console.log('No components to store');
-    return;
-  }
-  if (cacheStore.components.length == 0) {
-    cacheStore.components.push('0x0');
-  }
-  for (const component of components) {
-    var hexId = Uint8ArrayToHexString(component.id);
-
-    //CHECK INDEX
-    if (component.idx != cacheStore.components.length) {
-      console.log(
-        `Component.IDX ${component.idx} does not mach tail of list${cacheStore.components.length}`
-      );
-      continue;
-    }
-
-    cacheStore.components.push(hexId);
-    cacheStore.componentToIndex.set(hexId, component.idx);
-  }
-  console.log('Stored components: ' + cacheStore.components.length);
-}
-
-export function storeEntities(cacheStore: CacheStore, entities: Entity[]) {
-  if (typeof entities === 'undefined') {
-    console.log('No components to store');
-    return;
-  }
-  if (cacheStore.entities.length == 0) cacheStore.entities.push('0x0');
-  for (const entity of entities) {
-    var hexId = Uint8ArrayToHexString(entity.id);
-
-    //CHECK INDEX
-    if (entity.idx != cacheStore.entities.length) {
-      console.log(
-        `Entity.IDX ${entity.idx} does not match tail of list ${cacheStore.entities.length}`
-      );
-      continue;
-    }
-
-    cacheStore.entities.push(hexId);
-    cacheStore.entityToIndex.set(hexId, entity.idx);
-  }
-  console.log('Stored entities');
-}
-
-export async function storeValues(
-  cacheStore: CacheStore,
-  values: State[],
-  decode: ReturnType<typeof createDecode>
-) {
-  for (const event of values) {
-    const { packedIdx, data } = event;
-    let componentIdx = unpackTuple(packedIdx)[0];
-    const value = await decode(cacheStore.components[componentIdx], data);
-    cacheStore.state.set(packedIdx, value);
-  }
-  console.log('Stored values');
-}
-
-export function removeValues(cacheStore: CacheStore, values: State[]) {
-  for (const event of values) {
-    const { packedIdx, data } = event;
-    cacheStore.state.delete(packedIdx);
-  }
-  console.log('Removed values');
-}
-/************************************* END KAMIGAZE INTEGRATION *************************************/
 /**
  * Load from the remote snapshot service in chunks via a stream.
  *
@@ -259,7 +44,7 @@ export async function fetchSnapshotChunked(
   setPercentage?: (percentage: number) => void,
   pruneOptions?: { playerAddress: string; hashedComponentId: string }
 ): Promise<CacheStore> {
-  const cacheStore = createCacheStore();
+  const cacheStore = createStateCache();
   const chunkPercentage = Math.ceil(100 / numChunks);
 
   try {
@@ -287,14 +72,6 @@ export async function fetchSnapshotChunked(
   return cacheStore;
 }
 
-function Uint8ArrayToHexString(data: Uint8Array): string {
-  if (data.length === 0) return '0x00';
-  let hex = data.reduce((str, byte) => str + byte.toString(16).padStart(2, '0'), '');
-  if (hex.substring(0, 2) == '0x') hex = hex.substring(2);
-  const prefix = hex.length % 2 !== 0 ? '0x0' : '0x';
-  return prefix + hex;
-}
-
 /**
  * Reduces a snapshot response by storing corresponding ECS events into the cache store.
  *
@@ -309,7 +86,7 @@ export async function reduceFetchedState(
   decode: ReturnType<typeof createDecode>
 ): Promise<void> {
   const { state, blockNumber, stateComponents, stateEntities } = response;
-  const stateEntitiesHex = stateEntities.map((e) => Uint8ArrayToHexString(e) as EntityID);
+  const stateEntitiesHex = stateEntities.map((e) => uint8ArrayToHexString(e) as EntityID);
   const stateComponentsHex = stateComponents.map((e) => to256BitString(e));
 
   for (const { componentIdIdx, entityIdIdx, value: rawValue } of state) {
@@ -557,7 +334,7 @@ function createBlockCache(provider: JsonRpcProvider) {
 //   interval = 50,
 //   setPercentage?: (percentage: number) => void
 // ): Promise<CacheStore> {
-//   const cacheStore = createCacheStore();
+//   const cacheStore = createStateCache();
 
 //   const events = await fetchEventsInBlockRangeChunked(
 //     fetchWorldEvents,
@@ -571,46 +348,6 @@ function createBlockCache(provider: JsonRpcProvider) {
 
 //   return cacheStore;
 // }
-
-/**
- * Create a function to decode raw component values.
- * Fetches component schemas from the contracts and caches them.
- *
- * @param worldConfig Contract address and interface of the World contract
- * @param provider ethers JsonRpcProvider
- * @returns Function to decode raw component values using their contract component id
- */
-export function createDecode() {
-  const decoders: { [key: string]: (data: BytesLike) => ComponentValue } = {};
-  // hardcode world.component.components and world.component.systems to use uint256 schema
-  // NOTE: maybe compute these keys with keccaks or keep a constants file for readability
-  decoders['0x4350dba81aa91e31664a09d24a668f006169a11b3d962b7557aed362d3252aec'] = createDecoder(
-    ['value'],
-    [13]
-  ); // world.component.components
-  decoders['0x017c816a964927a00e050edd780dcf113ca2756dfa9e9fda94a05c140d9317b0'] = createDecoder(
-    ['value'],
-    [13]
-  ); // world.component.systems
-  async function decode(componentId: string, data: BytesLike): Promise<ComponentValue> {
-    if (!decoders[componentId]) {
-      debug('Creating decoder for', componentId);
-      const compID = componentId as keyof typeof ComponentsSchema;
-      let schema = ComponentsSchema[compID];
-      if (!schema) {
-        console.warn(`No schema found for component ${String(compID)}`);
-        // set bool as a default schema - only to prevent errors
-        // TODO: whitelist components to listen to (need in snapshot & here)
-        schema = { keys: ['value'], values: [0] };
-      }
-      decoders[componentId] = createDecoder(schema.keys, schema.values);
-    }
-    // Decode the raw value
-    return decoders[componentId]!(data);
-  }
-
-  return decode;
-}
 
 export function groupByTxHash(events: NetworkComponentUpdate[]) {
   return events.reduce(
